--- src/main/kotlin/com/example/minicraftserver/domain/character/entity/repository/CharacterRepository.kt
@@ -1,7 +1,0 @@
-package com.example.minicraftserver.domain.character.entity.repository
-
-import org.springframework.data.repository.CrudRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface CharacterRepository : CrudRepository<Character, Long>
