+++ src/main/kotlin/com/example/minicraftserver/domain/item/domain/repository/ItemRepository.kt
@@ -1,0 +1,8 @@
+package com.example.minicraftserver.domain.item.domain.repository
+
+import com.example.minicraftserver.domain.item.domain.Item
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ItemRepository : CrudRepository<Item, Long>
