+++ src/main/kotlin/com/example/minicraftserver/domain/user/domain/User.kt
@@ -1,4 +1,4 @@
-package com.example.minicraftserver.domain.user.entity
+package com.example.minicraftserver.domain.user.domain
 
 import com.example.minicraftserver.global.entity.BaseIdEntity
 import jakarta.persistence.Column
@@ -9,7 +9,6 @@ import jakarta.validation.constraints.NotNull
 @Table(name = "tbl_user")
 @Entity
 class User (
-
     id: Long,
 
     @NotNull
