+++ src/main/kotlin/com/example/minicraftserver/domain/work/domain/Work.kt
@@ -1,7 +1,7 @@
-package com.example.minicraftserver.domain.work.entity
+package com.example.minicraftserver.domain.work.domain
 
-import com.example.minicraftserver.domain.work.entity.enums.RegionType
-import com.example.minicraftserver.domain.work.entity.enums.WorkType
+import com.example.minicraftserver.domain.work.domain.enums.RegionType
+import com.example.minicraftserver.domain.work.domain.enums.WorkType
 import com.example.minicraftserver.global.entity.BaseIdEntity
 import com.fasterxml.jackson.annotation.JsonFormat
 import jakarta.persistence.Column
@@ -13,7 +13,6 @@ import java.time.LocalDateTime
 @Table(name = "tbl_work")
 @Entity
 class Work (
-
     id: Long,
 
     @NotNull
@@ -27,5 +26,5 @@ class Work (
     val startTime: LocalDateTime,
 
     @Column(columnDefinition = "INT DEFAULT 0")
-    val duration: Int
+    val duration: Int,
 ) : BaseIdEntity(id)
