+++ src/main/kotlin/com/example/minicraftserver/domain/work/domain/enums/RegionType.kt
@@ -1,0 +1,6 @@
+package com.example.minicraftserver.domain.work.domain.enums
+
+enum class RegionType {
+    MINE,
+    PASTURE
+}
