+++ src/main/kotlin/com/example/minicraftserver/domain/work/domain/enums/WorkType.kt
@@ -1,0 +1,6 @@
+package com.example.minicraftserver.domain.work.domain.enums
+
+enum class WorkType {
+    WAR,
+    COLLECTION
+}
