--- src/main/kotlin/com/example/minicraftserver/domain/work/entity/enums/RegionType.kt
@@ -1,7 +1,0 @@
-package com.example.minicraftserver.domain.work.entity.enums
-
-enum class RegionType {
-
-    MINE,
-    PASTURE
-}
