--- src/main/kotlin/com/example/minicraftserver/domain/work/entity/enums/WorkType.kt
@@ -1,7 +1,0 @@
-package com.example.minicraftserver.domain.work.entity.enums
-
-enum class WorkType {
-
-    WAR,
-    COLLECTION
-}
