+++ src/main/kotlin/com/example/minicraftserver/global/entity/BaseIdEntity.kt
@@ -10,7 +10,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener::class)
 abstract class BaseIdEntity (
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long,
