+++ src/main/kotlin/com/example/minicraftserver/global/error/GlobalExceptionHandler.kt
