+++ src/main/kotlin/com/example/minicraftserver/global/exception/InternalServerErrorException.kt
